Extract the Identity module (VerifiedIdentity keystone) from the flat kernel
Fourth #790 step (the payoff of the dependency inversion): now that VerifiedIdentity
and ValidatedLogin no longer depend on the OpenID/SAML modules (#796), move them out
of the flat Api kernel into Jellyfin.Plugin.SSO_Auth.Api.Identity.

- VerifiedIdentity, ValidatedLogin -> Api/Identity/ (namespace Api.Identity), import
  added at each call site. The keystone's allowed deps are Authz (permission grants)
  and Provider (link mode) — registered as the Identity -> {Authz, Provider} edge on
  the module DAG. Oidc/Saml/Flows now import Identity (protocols point INTO the
  keystone), and no lower module imports Identity (no cycle).

Pure mechanical move, no behavior change. Verified by the full build (--warnaserror,
both TFMs) and the module DAG + keystone-construction conformance rules; CI runs the
full suite on Linux (some SSOControllerAuthorizationTests fail LOCALLY on a Windows
application-control policy that blocks the in-process Kestrel test server's sockets —
a host issue, not this move, which is namespace-only; the baseline on origin/main is
green and the change touches no networking).

Refs #790
